sample_project/main/serializers.py
# sample_project/main/serializers.py

from rest_framework import serializers
from .models import City, Address


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = '__all__'


class AddressSerializer(serializers.ModelSerializer):
    class Meta:
        model = Address
        fields = '__all__'
